[AIRLFOW-YYY] Move pool commands to separate file
